refactoring
introduced a board-class to handle all stuff related to playing field.
That contains drawing the board, checking several win conditions and
making plays on it. It also provides ways to copy the current board into
a new one while returning a new instance and not only a reference to the
already existing one.

the bots now handles calculating the best for himself
the human only handles the user interaction
player handles stuff that is common for both bot and human
program handles flow of the program